Move new types into modules
